requirements.txt
tqdm
spacy
gensim
python-dotenv
torch~=2.10.0
numpy~=2.4.2
opencv-python~=4.13.0.92
scikit-learn
sentence-transformers
yt-dlp
mediapipe